refactor(elaborator): rename Resolver struct and crate path
Phase A of the resolver → elaborator rename. This commit covers only
the mechanical Rust-level changes needed for the code to build:

- Move src/resolver/ → src/elaborator/ and src/resolver.rs → src/elaborator.rs
- Rename pub struct Resolver → pub struct Elaborator
- Update all impl<...> Resolver<...> blocks to Elaborator
- Update use super::Resolver and crate::resolver::* paths
- Update rustdoc links from [`Resolver::xxx`] / [`crate::resolver::...`]
  to the new names (these are doc-link identifiers, semantically tied to
  the struct rename)

Method names (resolve_all_modules, lower_tir_from_state), span names
(resolve/*), and prose-level doc comments ("the resolver", "resolver-
synthesized", etc.) are NOT changed in this commit. They follow in
Phase C / Phase D.

No behavior change. Build and unit tests pass.
