+++ wado-compiler/src/flat_package.rs
@@ -92,7 +92,7 @@ pub struct FlatPackage {
     /// here by [`crate::monomorphize::monomorphize`], which adds the
     /// instantiation layer once it has materialised the concrete
     /// trait-method instances.
-    pub trait_env: std::sync::Arc<crate::resolver::trait_env::TraitEnv>,
+    pub trait_env: std::sync::Arc<crate::elaborator::trait_env::TraitEnv>,
 }
 
 impl FlatPackage {
