+++ wado-compiler/src/monomorphize/func_inst.rs
@@ -8,7 +8,7 @@ use std::sync::Arc;
 use crate::hashmap::{IndexMap, IndexSet};
 use crate::module_source::ModuleSource;
 use crate::name::{LocalMethodName, MethodName, mangle_generic_name};
-use crate::resolver::trait_env::TraitEnv;
+use crate::elaborator::trait_env::TraitEnv;
 use crate::tir::{
     CallArg, FunctionKind, FunctionRef, InstantiationKey, MonomorphInfo, ResolvedType, TirBinaryOp,
     TirBlock, TirExpr, TirExprKind, TirFunction, TirLocal, TirModule, TirParam, TirPattern,
