+++ wado-compiler/src/monomorphize/state.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 use crate::hashmap::IndexMap;
 use crate::module_source::ModuleSource;
 use crate::name::{LocalMethodName, MethodName, mangle_generic_name};
-use crate::resolver::trait_env::TraitEnv;
+use crate::elaborator::trait_env::TraitEnv;
 use crate::tir::{InstantiationKey, ResolvedType, TypeId, TypeTable};
 
 /// Tracks struct monomorphization state
