+++ wado-compiler/src/nir.rs
@@ -164,7 +164,7 @@ impl CallArg {
 /// `()` is private to this module so no code outside `tir` can build one —
 /// this makes direct struct-literal construction of `NirExprKind::MethodCall`
 /// impossible and channels every resolver-side emission through the
-/// single checkpoint maintained in `Resolver::build_tir_method_call`,
+/// single checkpoint maintained in `Elaborator::build_tir_method_call`,
 /// which in turn guarantees arguments were typechecked against the
 /// callee's declared parameter types.
 ///
@@ -179,8 +179,8 @@ impl NirExprKind {
     /// Sole constructor of [`NirExprKind::MethodCall`].
     ///
     /// Callers are expected to have typechecked `args` against the callee's
-    /// declared parameter types before reaching here.  Resolver-side
-    /// constructions flow through `Resolver::build_tir_method_call`;
+    /// declared parameter types before reaching here.  Elaborator-side
+    /// constructions flow through `Elaborator::build_tir_method_call`;
     /// post-resolve rewriters thread already-checked NIR through this
     /// function too so the variant's `invariant` field stays coherent.
     pub(crate) fn method_call(
